Moved the dto's and enum's
