+++ src/wardrobe/garment-category.enum.ts
